+++ tfe/resource_tfe_workspace.go
@@ -32,7 +32,19 @@ func resourceTFEWorkspace() *schema.Resource {
 			},
 		},
 
-		CustomizeDiff: validateAgentExecution,
+		CustomizeDiff: func(c context.Context, d *schema.ResourceDiff, meta interface{}) error {
+			err := validateAgentExecution(c, d, meta)
+			if err != nil {
+				return err
+			}
+
+			err = validateRemoteState(c, d, meta)
+			if err != nil {
+				return err
+			}
+
+			return nil
+		},
 
 		Schema: map[string]*schema.Schema{
 			"name": {
@@ -88,6 +100,18 @@ func resourceTFEWorkspace() *schema.Resource {
 				Type:     schema.TypeBool,
 				Optional: true,
 				Default:  true,
+			},
+
+			"global_remote_state": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Computed: true,
+			},
+
+			"remote_state_consumer_ids": {
+				Type:     schema.TypeSet,
+				Optional: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 
 			"operations": {
@@ -163,12 +187,6 @@ func resourceTFEWorkspace() *schema.Resource {
 					},
 				},
 			},
-
-			"external_id": {
-				Type:       schema.TypeString,
-				Computed:   true,
-				Deprecated: "Use id instead. The external_id attribute will be removed in the future. See the CHANGELOG to learn more: https://github.com/hashicorp/terraform-provider-tfe/blob/v0.24.0/CHANGELOG.md",
-			},
 		},
 	}
 }
@@ -190,6 +208,12 @@ func resourceTFEWorkspaceCreate(d *schema.ResourceData, meta interface{}) error 
 		QueueAllRuns:        tfe.Bool(d.Get("queue_all_runs").(bool)),
 		SpeculativeEnabled:  tfe.Bool(d.Get("speculative_enabled").(bool)),
 		WorkingDirectory:    tfe.String(d.Get("working_directory").(string)),
+	}
+
+	// Send global_remote_state if it's set; otherwise, let it be computed.
+	globalRemoteState, ok := d.GetOkExists("global_remote_state")
+	if ok {
+		options.GlobalRemoteState = tfe.Bool(globalRemoteState.(bool))
 	}
 
 	if v, ok := d.GetOk("agent_pool_id"); ok && v.(string) != "" {
@@ -251,6 +275,18 @@ func resourceTFEWorkspaceCreate(d *schema.ResourceData, meta interface{}) error 
 		}
 	}
 
+	remoteStateConsumerIDs, ok := d.GetOk("remote_state_consumer_ids")
+	if ok && globalRemoteState.(bool) == false {
+		options := tfe.WorkspaceAddRemoteStateConsumersOptions{}
+		for _, remoteStateConsumerID := range remoteStateConsumerIDs.(*schema.Set).List() {
+			options.Workspaces = append(options.Workspaces, &tfe.Workspace{ID: remoteStateConsumerID.(string)})
+		}
+		err = tfeClient.Workspaces.AddRemoteStateConsumers(ctx, workspace.ID, options)
+		if err != nil {
+			return fmt.Errorf("Error adding remote state consumers to workspace %s: %v", name, err)
+		}
+	}
+
 	return resourceTFEWorkspaceRead(d, meta)
 }
 
@@ -275,6 +311,7 @@ func resourceTFEWorkspaceRead(d *schema.ResourceData, meta interface{}) error {
 	d.Set("auto_apply", workspace.AutoApply)
 	d.Set("description", workspace.Description)
 	d.Set("file_triggers_enabled", workspace.FileTriggersEnabled)
+	d.Set("global_remote_state", workspace.GlobalRemoteState)
 	d.Set("operations", workspace.Operations)
 	d.Set("execution_mode", workspace.ExecutionMode)
 	d.Set("queue_all_runs", workspace.QueueAllRuns)
@@ -282,8 +319,6 @@ func resourceTFEWorkspaceRead(d *schema.ResourceData, meta interface{}) error {
 	d.Set("terraform_version", workspace.TerraformVersion)
 	d.Set("trigger_prefixes", workspace.TriggerPrefixes)
 	d.Set("working_directory", workspace.WorkingDirectory)
-	// TODO: remove when external_id is removed
-	d.Set("external_id", workspace.ID)
 	d.Set("organization", workspace.Organization.Name)
 
 	var sshKeyID string
@@ -311,6 +346,19 @@ func resourceTFEWorkspaceRead(d *schema.ResourceData, meta interface{}) error {
 
 	d.Set("vcs_repo", vcsRepo)
 
+	var remoteStateConsumerIDs []interface{}
+	if !workspace.GlobalRemoteState {
+		workspaceList, err := tfeClient.Workspaces.RemoteStateConsumers(ctx, id)
+		if err != nil {
+			return fmt.Errorf(
+				"Error reading remote state consumers workspace %s: %v", id, err)
+		}
+		for _, remoteStateConsumer := range workspaceList.Items {
+			remoteStateConsumerIDs = append(remoteStateConsumerIDs, remoteStateConsumer.ID)
+		}
+	}
+	d.Set("remote_state_consumer_ids", remoteStateConsumerIDs)
+
 	return nil
 }
 
@@ -323,7 +371,8 @@ func resourceTFEWorkspaceUpdate(d *schema.ResourceData, meta interface{}) error 
 		d.HasChange("file_triggers_enabled") || d.HasChange("trigger_prefixes") ||
 		d.HasChange("allow_destroy_plan") || d.HasChange("speculative_enabled") ||
 		d.HasChange("operations") || d.HasChange("execution_mode") ||
-		d.HasChange("description") || d.HasChange("agent_pool_id") {
+		d.HasChange("description") || d.HasChange("agent_pool_id") ||
+		d.HasChange("global_remote_state") {
 
 		// Create a new options struct.
 		options := tfe.WorkspaceUpdateOptions{
@@ -332,6 +381,7 @@ func resourceTFEWorkspaceUpdate(d *schema.ResourceData, meta interface{}) error 
 			AutoApply:           tfe.Bool(d.Get("auto_apply").(bool)),
 			Description:         tfe.String(d.Get("description").(string)),
 			FileTriggersEnabled: tfe.Bool(d.Get("file_triggers_enabled").(bool)),
+			GlobalRemoteState:   tfe.Bool(d.Get("global_remote_state").(bool)),
 			QueueAllRuns:        tfe.Bool(d.Get("queue_all_runs").(bool)),
 			SpeculativeEnabled:  tfe.Bool(d.Get("speculative_enabled").(bool)),
 			WorkingDirectory:    tfe.String(d.Get("working_directory").(string)),
@@ -428,6 +478,46 @@ func resourceTFEWorkspaceUpdate(d *schema.ResourceData, meta interface{}) error 
 		}
 	}
 
+	globalRemoteState := d.Get("global_remote_state").(bool)
+	if !globalRemoteState && d.HasChange("remote_state_consumer_ids") {
+		oldWorkspaceIDValues, newWorkspaceIDValues := d.GetChange("remote_state_consumer_ids")
+		newWorkspaceIDsSet := newWorkspaceIDValues.(*schema.Set)
+		oldWorkspaceIDsSet := oldWorkspaceIDValues.(*schema.Set)
+
+		newWorkspaceIDs := newWorkspaceIDsSet.Difference(oldWorkspaceIDsSet)
+		oldWorkspaceIDs := oldWorkspaceIDsSet.Difference(newWorkspaceIDsSet)
+
+		// First add the new consumerss
+		if newWorkspaceIDs.Len() > 0 {
+			options := tfe.WorkspaceAddRemoteStateConsumersOptions{}
+
+			for _, workspaceID := range newWorkspaceIDs.List() {
+				options.Workspaces = append(options.Workspaces, &tfe.Workspace{ID: workspaceID.(string)})
+			}
+
+			log.Printf("[DEBUG] Adding remote state consumers to workspace: %s", d.Id())
+			err := tfeClient.Workspaces.AddRemoteStateConsumers(ctx, d.Id(), options)
+			if err != nil {
+				return fmt.Errorf("Error adding remote state consumers to workspace %s: %v", d.Id(), err)
+			}
+		}
+
+		// Then remove all the old consumers.
+		if oldWorkspaceIDs.Len() > 0 {
+			options := tfe.WorkspaceRemoveRemoteStateConsumersOptions{}
+
+			for _, workspaceID := range oldWorkspaceIDs.List() {
+				options.Workspaces = append(options.Workspaces, &tfe.Workspace{ID: workspaceID.(string)})
+			}
+
+			log.Printf("[DEBUG] Removing remote state consumers from workspace: %s", d.Id())
+			err := tfeClient.Workspaces.RemoveRemoteStateConsumers(ctx, d.Id(), options)
+			if err != nil {
+				return fmt.Errorf("Error removing remote state consumers from workspace %s: %v", d.Id(), err)
+			}
+		}
+	}
+
 	return resourceTFEWorkspaceRead(d, meta)
 }
 
@@ -463,6 +553,23 @@ func validateAgentExecution(_ context.Context, d *schema.ResourceDiff, meta inte
 		d.SetNewComputed("operations")
 	} else if d.HasChange("operations") {
 		d.SetNewComputed("execution_mode")
+	}
+
+	return nil
+}
+
+func validateRemoteState(_ context.Context, d *schema.ResourceDiff, meta interface{}) error {
+	// If remote state consumers aren't set, the global setting can be either value and it
+	// doesn't matter.
+	_, ok := d.GetOk("remote_state_consumer_ids")
+	if !ok {
+		return nil
+	}
+
+	if globalRemoteState, ok := d.GetOk("global_remote_state"); ok {
+		if globalRemoteState.(bool) == true {
+			return fmt.Errorf("global_remote_state must be 'false' when setting remote_state_consumer_ids")
+		}
 	}
 
 	return nil